sms-code.bak/backend/requirements.txt
# requirements.txt

amqp==5.2.0
asarPy==1.0.1
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.3
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-filter==23.5
django-grappelli==3.0.8
django-rq==2.10.1
django-task==2.0.6
django-timezone-field==6.1.0
django4-tabbed-admin==1.0.6
djangorestframework==3.14.0
djangorestframework-jwt-5==1.13.0
future==0.18.3
iso8601==2.1.0
kombu==5.3.5
Markdown==3.5.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==1.7.1
pyserial==3.5
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
rq==1.15.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
